requirements.txt
flask
flask-login
flask-bcrypt
flask-sqlalchemy
requests
pillow
